+++ eslint-bridge/src/utils/utils-global.ts
